move size fn. to array package
